logic for correct-incorrect answers
