+++ src/components/Counter/counter.css
@@ -1,3 +1,9 @@
 .counter-container{
     color: white;
+    width: 100%;
+    margin: 0 auto;
+    display: flex;
+    justify-content: center;
+    background: rgb(30, 75, 158);
+      border-radius: 20px 20px 0 0 ;
 }
