+++ src/components/Score/score.css
@@ -1,14 +1,36 @@
 .score_container{
-    background: cyan;
     width: 60%;
+    margin: 10px auto;
+    display: flex;
+    justify-content: center;
+    width: 80%;
+    flex-direction: column;
+    align-items: center;
+    padding: 10px;
+    border: 1px solid rgb(163, 193, 196);
+}
+
+.score_text_container{
+    width: 100%;
     margin: 0 auto;
+    display: flex;
+    flex-direction: column;
+    padding: 10px 0;
+    border-bottom: 1px solid rgb(163, 193, 196);
+}
+
+.score_title{
+    margin: 10px auto;
+}
+
+.score_counter{
+   margin: 0 auto;
 }
 
 .score_map{
-    background: darkmagenta;
     display: flex;
     justify-content: center;
-    margin: 0 auto;
+    margin: 10px auto;
     align-items: center;
     flex-wrap: wrap;
 }
@@ -16,17 +38,18 @@
 .score_round{
     background:darkkhaki;
     border-radius: 100%;
-    width: 10px;
-    height: 10px;
+    width: 20px;
+    height: 20px;
     padding: 5px;
-    margin: 10px;
+    margin: 7px;
 }
 
 .score_question_number{
     color: white;
     align-items: center;
+    text-align: center;
     margin: 0 auto;
-    font-size: 0.5rem;
+    font-size: 0.7rem;
 }
 
 .correct{
