+++ src/components/Timer/index.js
@@ -1,0 +1,21 @@
+{ /* <form className="form_radio" onSubmit={(event) => event.preventDefault()}>
+          {question.options.map((option, index) => {
+            return (
+              <label
+                className="form_radio_label"
+                htmlFor="input_radio"
+                key={index + 1}>
+                <input
+                  disabled={answers[currentQuestionIndex]}
+                  className="form_radio_input"
+                  id="input_radio"
+                  required
+                  type="radio"
+                  value={option}
+                  name={question.id}
+                  onChange={() => handleOnChange(index)} />
+                {option}
+              </label>
+            )
+          })}
+        </form> */ }
