+++ src/index.css
@@ -7,15 +7,28 @@ body {
   -moz-osx-font-smoothing: grayscale;
   background: rgb(2,0,36);
   background: linear-gradient(150deg, rgba(2,0,36,1) 0%, rgba(152,121,173,1) 100%, rgba(0,212,255,1) 100%);
-  min-height: 100vh;
   display: flex;
-  text-align: center;
   justify-content: center;
-  align-items: center;
-  
+  margin: 0;
+  padding: 0;
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
+}
+
+.root{
+margin: 0;
+  display: flex;
+  flex-direction: column;
+}
+
+@media (min-width: 768px) {
+.root{
+  flex-direction: row;
+  width: 90%;
+  gap: 0.5rem;
+}
+  
 }
